requirements.txt
a==1.0
aerich==0.6.3
aiofiles==0.8.0
aiomysql==0.1.1
aiosqlite==0.17.0
aniso8601==9.0.1
anyio==3.6.1
asgiref==3.5.2
asyncodbc==0.1.1
asyncpg==0.26.0
atlastk==0.13.2
attrs==22.1.0
Automat==20.2.0
azure-core==1.24.2
azure-storage-blob==12.13.0
bcrypt==3.2.2
cached-property==1.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
cheroot==8.6.0
CherryPy==18.8.0
click==8.1.3
constantly==15.1.0
cryptography==37.0.4
cssselect==1.1.0
cx-Oracle==8.3.0
ddlparse==1.10.0
decorator==5.1.1
dictdiffer==0.9.0
distlib==0.3.5
dnspython==2.2.1
email-validator==1.2.1
et-xmlfile==1.1.0
fastapi==0.79.0
fastapi-contrib==0.2.11
fastapi-login==1.8.2
fastapi-pagination==0.9.3
filelock==3.7.1
graphene==3.1
graphene-file-upload==1.3.0
graphql-core==3.2.1
graphql-relay==3.2.0
greenlet==1.1.2
h11==0.12.0
h2==4.1.0
hpack==4.0.0
httpcore==0.15.0
httptools==0.4.0
httpx==0.23.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.3
importlib-resources==5.9.0
incremental==21.3.0
install==1.3.5
iso8601==1.0.2
isodate==0.6.1
itemadapter==0.6.0
itemloaders==1.0.4
itsdangerous==2.1.2
jaraco.classes==3.2.2
jaraco.collections==3.5.2
jaraco.context==4.1.2
jaraco.functools==3.5.1
jaraco.text==3.8.1
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.9.0
lxml==4.9.1
MarkupSafe==2.1.1
minio==7.1.15
more-itertools==8.13.0
motor==3.0.0
msrest==0.7.1
mtaa==1.5
numpy==1.23.1
oauthlib==3.2.0
of==1.0.1
openpyxl==3.0.10
oracledb==1.0.2
pandas==1.4.3
parsel==1.6.0
passlib==1.7.4
pendulum==2.1.2
phonenumbers==8.12.52
pkgutil_resolve_name==1.3.10
portend==3.1.0
priority==2.0.0
promise==2.3
Protego==0.2.1
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.9.1
PyDispatcher==2.0.5
Pygments==2.12.0
PyJWT==2.4.0
pymongo==4.2.0
PyMySQL==1.0.2
pyodbc==4.0.34
pyOpenSSL==22.0.0
pyparsing==3.0.9
pypika-tortoise==0.1.6
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2022.1
pytzdata==2020.1
PyYAML==6.0
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
Rx==3.2.0
Scrapy==2.6.2
sentinel==1.0.0
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
SQLAlchemy-Utils==0.38.3
starlette==0.19.1
starlette-core==0.0.1
starlette-graphene3==0.6.0
Starlette-WTF==0.4.3
syncer==1.3.0
tempora==5.0.2
tldextract==3.3.1
to==0.3
tomlkit==0.11.1
tortoise-orm==0.19.2
Twisted==22.4.0
typing==3.7.4.3
typing_extensions==4.3.0
ujson==5.4.0
urllib3==1.26.11
uvicorn==0.18.2
uvloop==0.16.0
w3lib==1.22.0
watchdog==3.0.0
watchgod==0.8.2
websockets==10.3
ws4py==0.5.1
wsproto==1.1.0
WTForms==3.0.1
xlrd==2.0.1
XlsxWriter==3.0.3
xmltodict==0.13.0
zc.lockfile==2.0
zipp==3.8.1
zope.interface==5.4.0
